Adding a lot of casts to support ComputeCpp 0.8.0, a lot of these deal with ambigious calls using overloaded vec operators. This version of ComputeCpp is a lot more consistent in returning swizzled vec types even when returning a single element. The methods I found to get around this are to cast the vec type to the underlying type for example: (((float)vec3.x()) >= 0.0f) or to do a call as follows: (vec3.x() >= 0.0f).x(). The dataT() operator doesn't appear to be supported on the host yet, hence the use of .x() twice, the result of the >= operator is a swizzled vec that we then retrieve the only element of. Now, the problem with using the member call over the casting is that it's no longer backwards compatible. Older versions of ComputeCpp don't return swizzled vecs as a result of these operations so its a compiler error. At the moment I've chosen to cast in scenarios where the calls are ambigious or result in an incorrect return type to maintain compatability with older versions.
Minor error in readme also fixed.

+++ kfusion/include/default_parameters.h
@@ -467,8 +467,8 @@ struct Configuration {
 				std::cerr << "update map_size to " << this->volume_size.x()
 						<< "mx" << this->volume_size.y() << "mx"
 						<< this->volume_size.z() << "m" << std::endl;
-				if ((this->volume_size.x() <= 0) || (this->volume_size.y() <= 0)
-						|| (this->volume_size.z() <= 0)) {
+				if ((((float)this->volume_size.x()) <= 0.0f) || (((float)this->volume_size.y()) <= 0.0f)
+						|| (((float)this->volume_size.z()) <= 0.0f)) {
 #else
 				std::cerr << "update map_size to " << this->volume_size.x
 						<< "mx" << this->volume_size.y << "mx"
@@ -499,9 +499,9 @@ struct Configuration {
 						<< this->volume_resolution.x() << "x"
 						<< this->volume_resolution.y() << "x"
 						<< this->volume_resolution.z() << std::endl;
-				if ((this->volume_resolution.x() <= 0)
-						|| (this->volume_resolution.y() <= 0)
-						|| (this->volume_resolution.z() <= 0)) {
+				if ((((unsigned int)this->volume_resolution.x()) <= static_cast<unsigned int>(0))
+						|| (((unsigned int)this->volume_resolution.y()) <= static_cast<unsigned int>(0))
+						|| (((unsigned int)this->volume_resolution.z()) <= static_cast<unsigned int>(0))) {
 #else
 						<< this->volume_resolution.x << "x"
 						<< this->volume_resolution.y << "x"
